+++ src/schick/seg073.h
@@ -3,11 +3,11 @@ namespace M302de {
 #endif
 
 /* 0x20 */
-unsigned short get_tavern_gossip(void);
+signed int get_tavern_gossip(void);
 /* 0x25 */
 char* get_drinkmate(void);
 /* 0x2a */
-signed short tavern_quest_infos(void);
+signed int tavern_quest_infos(void);
 /* 0x2f */
 void tavern_follow_informer(void);
